Added '%access' directive and updated libs
%access sets the 'default' accessibility to public, private or abstract.
If there are *no* access modifiers in a file, everything is public
(this is the same behaviour as a Haskell module with no export list).
If there are *any* access modifiers in a file, then declarations with
no accessibility modifier take the %access setting as default.
The default %access setting is 'private'.