Disentangle Haskeline from the internals Idris
Now, only the areas that actually work with the user involve the InputT monad
transformer. The practical advantage of this is that different Haskeline
settings structures can be used for different shells - the main REPL, the
prover, and the TT shell, simply by providing a different completion function
in the runInputT call.