Remove the unicode from the source files.
This has tripped up at least a couple people trying to install idris, because their
LANG setting did not include unicode, or something.  This comment was just copied
and pasted from the comments of the Haskell lib.  The benefits of having a math
symbol in there don't seem to be worth the pain it causes some people.