Add scripts
This commit adds a new command-line argument to Idris that allows an arbitrary
expression to be run in the executor. Two forms are supported:

idris Foo.idr --opt1 --exec

causes Main.main to be run after loading Foo.idr, while

idris Foo.idr --opt1 --exec expr

causes expr to be run using the executor.

This can be used to run Idris programs without having to compile them, giving
a faster startup for quick tasks. It can also be used to test the behavior of
the executor against the behavior of compiled code more easily.