Switch on 'elabPE'
Generates a new specialised definition for [static] argument positions
which are then elaborated recursively.

Not yet transforming from the original call to the specialised
definition. Also still need to take more care in how we specialise
definitions - however, the mechanics of generating the new definitions
is now done.

Generating new definitions may fail for various reasons (scoping errors,
for example). This is considered safe, and just fails silently, because
it just means it's not a specialisable call.

Also removed the {static} syntax because it's inconsistent. [static]
only works on named pi bindings due to awkwardness in the parser.