Don't reset totality for non-total things
We recheck everything at the end of checking a module, but if things
are currently known non-covering, we're not going to check that again,
so resetting means that that information is lost. So better only to
recheck things whcih are currently believed total, now that we have
complete information about the module.

Fixes #689