Update llvm-general and llvm-general-pure deps.
I originally was trying to be 'clever' and ensure Idris could handle
both llvm 3.3.8.* and 3.4.*. This wasn't really working so I've opted
for simply 3.4.* dependencies to match the latest version of the LLVM
packages on Hackage.

There are some non-backwards compatible changes for this and I've tried
to fix some of the obvious ones but I might have missed some more subtle ones.