Add --testpkg option + tests section in IPKG files
Add a new section to IPKG files called "tests". This section should be a
comma-separated list of fully qualified names of IO actions, the
sequence of which in some way test an Idris package.

When invoked as idris --testpkg IPKG, Idris will generate a module in a
temporary directory that imports every module from the package IPKG and
whose main function simply executes each specified test.

The intention is to keep test specifications as low-level as possible,
allowing user libraries to define a better mechanism on top of this.