Added rts/arudino/idris_main.c. It is similar to the standard idris_main.c except it does not attempt to handle argv/argc and it also calls the Arduino init() function.
Additionally, all the rts .c files are now included in the data-files section of the idris.cabal file. When using Idris to cross-compile, we need to recompile the RTS for the target platform. We do not know what platforms might be targeted in advance, so we need to have all the RTS source files available at compile time.