Invoke external code generator
If --codegen specifies somethng which isn't one of the built-in code
generators, invoke an external CG. For --codegen foo, this should be a
command which takes the following form:

idris-foo <ibc-file-list> [-o <output-file>]

idris-foo should read the ibc files, and generate an output file called
<output-file>. Naturally it's easiest to use Idris as a library to do
this. See the codegen/ directory in the Idris-dev repository for
examples.