Search in current elaboration stack for instances
That is, keep a stack of what is currently being elaborated (parents of
where clauses, case blocks, etc) so that the most local type class
instance is always chosen when resolving. In particular, this means
when defining named instances, the instance being defined is always used
rather than the global instance.

Fixes #871