Allow implicits to appear anywhere
This is still limited: in particular, there are no implicit lambdas and
you can't give values of scoped implicits explicitl, unlike top level
implicits.

Basically, if the elaborator sees it's making something of type
({x : _} -> y} it will wrap an implicit lambda around it.
If it sees any scoped implicits in the name it's
elaborating, it will add an implicit _.