Use a proper conversion check for type provider types
Previously, the type check used a syntactic equality. Probably, this
should be rewritten as a tactic script instead of the current code, as I
fear that there might be type class resolution issues for effect
handlers in the ambiguous cases as things currently stand.

In any case, this makes broken code start working again.