Expose equality proofs in with rule
Given a pattern matching definition of the form:
foo args with (expr) proof q
  foo args | pat = rhs

In 'rhs', 'q' is in scope, and is a proof that pat = expr. This can
simplify some proofs (e.g. as in #1920)

I'm not sure about the syntax for this (and I haven't documented it
anyway), so this is EXPERIMENTAL for the moment. Syntax could change or
disappear...