Make the Makefile cabal-sandbox aware.
The Makefile has been augmented to construct a relative location to a
pre-installed Idris executable based upon whether a sandbox was
involved. This fix resolves problems when using `make relib` and
running testing commands that rely upon a preinstalled idris binary.

The Makefile checks for the existence of the directory: `.cabal-sandox`.
If the directory exists then
    `IDRIS := .cabal-sandbox/bin/idris`
else
    `IDRIS := dist/build/idris`