Make Quotable polymorphic on quotation type
Quotable now accepts an additional parameter: the goal type to quote
to. This allows it to be used for both Raw and TT quotations.
Unfortunately, the bodies of the instances are mostly copy-pasted, but I
couldn't find a way to re-use the code because the elaboration of the
quasiquotes is not at all the same.