Change to a recoverable error in proof search
This means if a constructor fails when deep in search, instead of
abandoning the whole thing we'll just move on to the next metavariable.
This effectively gives backtracking for the 'search' tactic.

Of course, care is needed in choosing what to search for (make sure the
types are descriptive enough to cut off search, in particular), because
if it goes really deep this could take a long time!