Support externally defined primitives
%extern name : type -- declares a function which is treated as an
operator, and compiled specially by a back end. This can be used in
conjuction with an abstract data declaration to add new primitives
specific to some back end.

Also made Ptr and ManagedPtr external primitives (they have no literals,
so no need for them to be part of the compiler)