Replace the Monoid instance on Maybe
The current default Monoid instance for Maybe is like that in Haskell,
where a semigroup operator on the underlying type is used to collect
successes. Conor argues convincingly that this is not the right notion
of computation for Maybe [1], so I figured I'd fix it while we're young.

[1] https://www.reddit.com/r/haskell/comments/30s1t2/proposal_make_semigroup_as_a_superclass_of_monoid/cpvdco1

The old behavior is available in a named instance, and both are documented.