Minor change to disambiguation rules
When trying alternatives, if elaborating the first introduces unresolved
type classes, consider this a failure and elaborate the second.
Normally, type classes should resolve as soon after application as
possible.

(aside: we need to write down precise rules for type-directed
disambiguation; they are not straightforward due to the possibility of
delaying type class resolution, auto implicits and unresolved
unification problems. There is a trade off between successful
disambiguation and efficient enough compilation, so we should write down
what the current trade off is.)