Improved type-directed disambiguation
If disambiguation fails, then instead of reporting an error immediately,
defer the elaboration until later. Fixes #2285.

In the process, PAlternative now has a more descriptive argument than
merely a Bool to say whether it's a proof search, or a disambiguation,
and whether it's allowed to be delayed or not. IBCs need rebuilding.