Environment fix in prepare_apply
One for the "How did this ever work" files. When claiming a
metavariable for a function argument, we need to take the current
environment before called 'forget' so that names are correct.
Fixes #2211

(Okay, I know how this worked before really: essentially it used an
assumption which is violated by scoped implicits, so now no longer make
that assumption.)