Replace absurd syntax rule FCs
Previously, FCs coming from syntax rules would keep the FCs of their
definition context. Now, after expanding a syntax rule, the FCs that
aren't contained in the region in which the syntax was expanded are
replaced by a reference to the present file.

This does not give a precise source location - that runs the risk of
generating false highlighting information - but this solution at least
places the error in the right file, and the "While elaborating ..." bit
can be used to find the error. We should probably get better at
separating the highlighting-relevant FCs from the
highlighting-irrelevant FCs in the future.