Trim trailing whitespace in parseCmd
When auto-completing a filename to :load from the REPL, the trailing
space added by the auto-completion will be ignored for actually loading
the file, but a subsequent invocation of :edit will use the path
including the space, leading to the creation of a new file
"<module>.idr .idr".

This problem is easily fixed by trimming not just leading, but also
trailing whitespace in parseCmd.