Annotate links with their URL
This allows frontends other than the console to display links with a
special decoration. For example, the Emacs mode underlines the link text
and makes it clickable, and the HTML output for :printdef uses the
appropriate tag. The console still shows the link text parenthesized
after the underlined link text.