Make 'case' more flexible with dependent types
Implicit variables remain implicit in the generated case (or, more
accurately, variables which are unused in the rhs are made implicit on
the lhs), meaning that it is no longer necessary for each branch to
target the same type or for each case of the scrutinee to have the same
type.

In other words, case is now allowed provided that it does not affect the
form of any variable which is explicitly used in the case block.

Example, which is now allowed but wasn't before:

  app : Vect n a -> Vect m a -> Vect (n + m) a
  app xs ys = case xs of
                   [] => ys
                   (x :: xs) => x :: app xs ys