Restrict (and speed up) implicit coercions
'implicit' will now only work if there is a concrete argument type and a
concrete return type. This means we can rule out impossible coercions
much quicker, and has the effect of speeding up error reporting by not
trying to fix up the error with coercions which are never going to work.

There's a special case for "lend" but UniqueType is experimental and
probably needs a rethink, so I'm okay with this at least for now.