Make interactive proof search less enthusiastic
Instead of guessing, it will only fill in a metavariable if all of the
available constructors/hints have disjoint types. This means you'll no
longer always get 'Nothing' for Maybe or '[]' for List, but rather
you'll get a new metavariable.

'auto' behaves as before, because the idea there is to find any old
thing with the right type.

It feels like interactive proof search should be a bit more refined,
though, because the idea of that is to find the program you really want,
rather than any old program...

More refinements to this are likely, for example refusing to fill in
with locals if there are clashing types.