Use subst for intro, rather than updsubst
This is because we are introducing things that aren't holes, and the
application status being "complete" might block for substitution of an
introduction. This fixes a bug in the interactive prover and the
interactive elaborator where introduction only worked when precisely the
same name was provided, becuase it was incompletely substituted.