Add docs, stack.yaml and benchmarks to idris.cabal
Added the text and Markdown documentation to sdists and installations
with the extra-doc-files cabal field. This misses all the Sphinx docs -
the effect of that field is just to copy the files as is and we really
want the outputted HTML. I'm sure it's possible to generate the HTML
with Sphinx at the same time as Haddock, I just didn't bother figuring
out how. I'll add an issue.

Added stack.yaml and the benchmarks to extra-source-files. This just
copies them in sdists.