Neg and Nat adjustments
Neg now means "Numbers with possible negative values" and the methods
which deal with negative numbrs (abs, - and negate) are now part of this
rather than Num.

This means, in particular, that - on Nat can now require a proof.
'minus' on Nat behaves as before (and the proofs use minus rather than
-).