Reduce space leak through ugly DeepSeq hack
The space leak documented in issue #2386 is helped by forcing the state
in a few places. Now, with the empty module, no space leak occurs, while
it grows by about 1 MB per 200 reloads with a large module. While
deepseqing the state is a hack, and should be better, I've not yet
improved on this and it does make Idris usable for repeated reloads of
large modules when using the interactive editing features.